+++ tests/attributes/mqt/test_demographics.py
@@ -10,28 +10,17 @@ from nacc_attribute_deriver.symbol_table import SymbolTable
 
 
 @pytest.fixture(scope="function")
-def table() -> SymbolTable:
+def table(uds_table) -> SymbolTable:
     """Create dummy data and return it in an attribute object."""
-    data = {
-        "file": {
-            "info": {
-                "forms": {
-                    "json": {  # needed for DemographicsAttributeCollection
-                        "sex": "1",
-                        "primlang": "2",
-                        "visitdate": "2025-01-01",
-                        "module": "uds",
-                        "packet": "I",
-                        "birthmo": 1,
-                        "birthyr": 1950,
-                        "formver": 3.0,
-                        "naccid": "NACC123456",
-                        "adcid": 0,
-                    }
-                }
-            }
-        },
-        "subject": {  # needed for DerivedDemographicsAttributeCollection
+    # needed for DemographicsAttributeCollection
+    uds_table['file.info.forms.json'].update({
+        "sex": "1",
+        "primlang": "2",
+    })
+
+    # needed for DerivedDemographicsAttributeCollection
+    uds_table.update({
+        "subject": {
             "info": {
                 "derived": {
                     "np_death_age": 80,
@@ -42,9 +31,9 @@ def table() -> SymbolTable:
                 }
             }
         },
-    }
+    })
 
-    return SymbolTable(data)
+    return uds_table
 
 
 class TestDemographicsAttributeCollection: